update abstract & interface
