+++ Assets/Scenes/Gameplay.unity
@@ -778,6 +778,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _speed: 2
+  _randomSpeed: 2
 --- !u!95 &451396451
 Animator:
   serializedVersion: 4
@@ -880,7 +881,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   scoreManager: {fileID: 1740509438}
   enemyManager: {fileID: 1658817869}
-  direction: 0
 --- !u!1 &459524802
 GameObject:
   m_ObjectHideFlags: 0
@@ -1234,6 +1234,51 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 589640151}
   m_CullTransparentMesh: 1
+--- !u!1 &624512911
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 624512913}
+  - component: {fileID: 624512912}
+  m_Layer: 0
+  m_Name: RaycastManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &624512912
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 624512911}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b66eaf3b49b73d746a707d82e041275d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  scoreManager: {fileID: 1740509438}
+--- !u!4 &624512913
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 624512911}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 8
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &974989526
 GameObject:
   m_ObjectHideFlags: 0
@@ -1512,7 +1557,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 090fc01a00a0cc84188e20b4510be9d1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _speed: 2.5
+  _speed: 3
+  _randomSpeed: 0
 --- !u!114 &1001432611
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1689,7 +1735,6 @@ GameObject:
   - component: {fileID: 1211009857}
   - component: {fileID: 1211009856}
   - component: {fileID: 1211009858}
-  - component: {fileID: 1211009859}
   m_Layer: 0
   m_Name: Goal
   m_TagString: Goal
@@ -1790,19 +1835,6 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 1}
   m_EdgeRadius: 0
---- !u!114 &1211009859
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1211009854}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b66eaf3b49b73d746a707d82e041275d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  rayRange: 4
 --- !u!1 &1273031402
 GameObject:
   m_ObjectHideFlags: 0
@@ -2633,6 +2665,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _speed: 2
+  _randomSpeed: 0
 --- !u!1 &1822364774
 GameObject:
   m_ObjectHideFlags: 0
