+++ Assets/Scripts/Command.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ea8ac672f09e144cb3832625726a077
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
