+++ Assets/Scripts/Enemy/Enemy.cs
@@ -1,36 +1,42 @@
+using Manager;
+using Raycast;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Enemy : MonoBehaviour, IRaycastable
+namespace Enemy
 {
-    //public float speed;
-    public ScoreManager scoreManager;
-    public EnemyManager enemyManager;
-
-    void Start()
+    public class Enemy : MonoBehaviour, IRaycastable
     {
-        
-    }
+        //public float speed;
+        public ScoreManager scoreManager;
+        public EnemyManager enemyManager;
 
-    void Update()
-    {
-        if (transform.position.y <= -5)
+        void Start()
         {
-            scoreManager.Life();
 
-            enemyManager.enemyDestroyed.Add(enemyManager.enemyList[0]);
+        }
+
+        void Update()
+        {
+            if (transform.position.y <= -5)
+            {
+                scoreManager.Life();
+
+                enemyManager.enemyDestroyed.Add(enemyManager.enemyList[0]);
+                Destroy(gameObject);
+            }
+        }
+
+        public void Interact()
+        {
+            scoreManager.getScore += 10;
+
             Destroy(gameObject);
+            enemyManager.enemyDestroyed.Add(enemyManager.enemyList[0]);
+            gameObject.SetActive(false);
+            Debug.Log("Enemy Destroy");
         }
     }
-
-    public void Interact()
-    {
-        scoreManager.getScore += 10;
-
-        Destroy(gameObject);
-        enemyManager.enemyDestroyed.Add(enemyManager.enemyList[0]);
-        gameObject.SetActive(false);
-        Debug.Log("Enemy Destroy");
-    }
 }
+
