+++ Assets/Scripts/Human/Human.cs
@@ -1,22 +1,28 @@
+using Manager;
+using Raycast;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Human : MonoBehaviour, IRaycastable
+namespace Human
 {
-    public ScoreManager scoreManager;
-
-    void Update()
+    public class Human : MonoBehaviour, IRaycastable
     {
-        if (transform.position.y <= -5)
+        public ScoreManager scoreManager;
+
+        void Update()
         {
-            scoreManager.getScore += 50;
-            Destroy(gameObject);
-            Debug.Log(transform.position);
+            if (transform.position.y <= -5)
+            {
+                scoreManager.getScore += 50;
+                Destroy(gameObject);
+                Debug.Log(transform.position);
+            }
+        }
+        public void Interact()
+        {
+            scoreManager.GameOver();
         }
     }
-    public void Interact()
-    {
-        Debug.Log("Interact");
-    }
 }
+
