+++ Assets/Scripts/Manager.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ab008fb9dda8a3241b1752e41e88becb
+guid: e8ab031d77221114191c9669f6cecdbe
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
