+++ Assets/Scripts/Manager/GameManager.cs.meta
