--- Assets/Scripts/ScenesManager.cs
@@ -1,28 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class ScenesManager : MonoBehaviour
-{
-    // Start is called before the first frame update
-    public void Awake()
-    {
-        Time.timeScale = 1;
-    }
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void ChangeScene(string scenename)
-    {
-        SceneManager.LoadScene(scenename);
-    }
-}
