--- Assets/Scripts/Useless/InputMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a4f5739b500f26e41b8d509c6a52977e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
