--- Assets/Scripts/WaveManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f1029005e0cca5d4c8f8e9d372708823
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
